+++ app/src/androidTest/java/woowacourse/payments/ui/CardOwnerNameTextFieldTest.kt
@@ -7,6 +7,7 @@ import androidx.compose.ui.test.performTextInput
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
+import woowacourse.payments.ui.component.CardOwnerNameTextField
 
 class CardOwnerNameTextFieldTest {
     @get:Rule
