+++ app/src/main/java/woowacourse/payments/ui/CardAdditionScreen.kt
@@ -6,6 +6,7 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
+import woowacourse.payments.ui.component.CardAdditionTopBar
 
 @Composable
 fun CardAdditionScreen(modifier: Modifier = Modifier) {
