+++ app/src/main/java/woowacourse/payments/ui/component/CardOwnerNameTextField.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui
+package woowacourse.payments.ui.component
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.material3.OutlinedTextField
